feat: display lighthouse runtime error when re-queuing failed routes
For me, an absent NODE_EXTRA_CA_CERTS environment variable caused the
lighthouse tasks to fail due to invalid TLS certificates after an
upgrade in a containerized environment. In that case, the existing error
message wasn't helpful, because the same URL worked fine on the host
where the certificate was correctly installed and picked up.